requirements.txt
streamlit
qrcode[pil]
pillow
qrcode
qreader